Fix zip-related bugs
The problem with zip is that it truncates to the shortest vector. My guess is that it is not very efficient for vectors either, as it makes a temporary vector, which surely hits the heap. Fix both problems by creating new functions foldr2 and foldr3, which behave like foldr except that they take two/three argument vectors.